==> doc/03-data-mapping.Rmd <==
---
title: "Data Mapping "
subtitle: "metalm"
author: "philippe.brunier@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      python =reticulate::eng_python,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))

require(dplyr)
require(kableExtra)
require(readr)
require(reticulate)
source('~/dev/metalm/R/function.R')
```

# File

I file utilizati in questo progetto sono riassunti in 3 tabelle principali: **database_opt.csv** con le estrazioni delle prove svolte al tornio. Tale file contiene i dati delle passate, espressi come media e deviazione standard sui tre assi cartesiani. **tabella_ox.txt** contiene le composizioni delle inclusioni rilevate al SEM, **tool_wear.txt** i dati di usura. Il contenuto metallografico, assieme alle prove meccaniche ed alle chimiche di prodotto, si trova invece in **metallurgical.csv** 


**tutti i file sono .csv con separatore '|'** 

## database_opt.csv
contiene i dati delle prove svolte in officina. Qui vengono registrare le informazioni per ogni passata, per ogni pezzo, per ogni prova. I valori delle pressioni, vengono ricavati tramite le seguenti formule:

$K = \frac{F}{area} [\frac{N}{mm^{2}}]$ 

dove 

$area = {feed}*{DoC} [mm^{2}]$

```{r caricamento_file_opt, include = TRUE}
header_opt <- read_delim('/data/metalm/header/head_opt.csv', delim=';', col_type=cols())
```

Il file è composto da `r nrow(header_opt)` colonne, con le seguenti specifiche



```{r plot_tabella_opt, include = TRUE}
cas_kable(header_opt, caption='optimization')
```


## tabella_ox.txt
contiene le composizioni delle inclusioni tipiche della marca espresse come frazione. Le inclusioni presenti sono importanti sia in termini di lavorabilità che in termini di processo in quanto correlabili con la fabbricazione AOD.

```{r caricamento_file_ox, include = TRUE}
header_ox <- read_delim('/data/metalm/header/head_ox.csv', delim=';', col_type=cols())
```

Il file è composto da `r nrow(header_ox)` colonne, con le seguenti specifiche

```{r plot_tabella_ox, include = TRUE}
cas_kable(header_ox, caption='oxides')
```

## metallurgical.csv
contiene i seguenti dati:

+ esiti prove meccaniche: trazione, durezza
+ esiti micropurezza
+ chimiche di prodotto

```{r caricamento_file_met, include = TRUE}
header_met <- read_delim('/data/metalm/header/head_met.csv', delim=';', col_type=cols())
```

Il file è composto da `r nrow(header_met)` colonne, con le seguenti specifiche

```{r plot_tabella_met, include = TRUE}
cas_kable(header_met, caption='metallurgical')
```


## tool_wear.txt
contiene i dati relativi alle prove di usura. Il tempo limite di contatto è definito come il picco sulla componente tangenziale (z)

```{r caricamento_file_wear, include = TRUE}
header_wear <- read_delim('/data/metalm/header/head_wear.csv', delim=';', col_type=cols())
```

Il file è composto da `r nrow(header_wear)` colonne, con le seguenti specifiche

```{r plot_tabella_wear, include = TRUE}
cas_kable(header_wear, caption='tool wear')
```


```{python prova, include = TRUE}
import pandas
dd = pandas.read_csv('/data/metalm/header/head_met.csv',sep=';')
nrows = len(dd.columns)
```
```{r converting, include = TRUE}
nrighe <- py$nrows
```
