==> doc/07-conclusioni.Rmd <==
---
title: "Data Discovery "
subtitle: "metalm"
author: "ffiou@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
require(dplyr)
require(rlang)
require(kableExtra)
require(readr)
require(ggplot2)
require(reticulate)
source('~/dev/metalm/R/function.R')
```


```{python import, include=FALSE}
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.table as tbl
import numpy as np
import ternary

plt.style.use('seaborn')

```

Definisco i colori dei materiali

```{python tabella_colori}

# Creazione dataframe colori materiali

colori_materiali =  pd.DataFrame ([['b'],['g'],['r']],
                    index=['F304L1','F304LO','F304LX'],
                    columns=['colore'])

```

# Box plot usura

leggiamo i dati

```{python lettura_wear}

#leggo i dati

filename = '/data/metalm/usura.pkl'

#leggiamo le colonne separate da |
df = pd.read_pickle(filename)

#togliamo LH dai dati
df = df.drop(df[df.materiale=='F304LH'].index)  

```

## Comparativa medie materiali

```{python boxplot_mat}

materiali = np.unique(df.materiale)

array = []

materialilabel = []

i=0

for materiale in materiali:
  
  colore = str(colori_materiali.loc[materiale,'colore'])
  
  df_materiale = df[df.materiale == materiale]
  
  #print(colore, materiale)
  
  myplot = df_materiale.boxplot(column='tempo_usura', 
  #patch_artist=True,
  #notch = True,
  color = colore,
  positions = [i],
  )
  
  array.append(i)
  
  materialilabel.append(materiale)
  
  i=i+1
  
  #print(array)

plt.xticks(array, materialilabel)
myplot.set_ylabel('tempo limite usura [min]')
myplot.set_title('')
myplot.set_xlabel('Materiali')
plt.show()
plt.savefig('/data/metalm/output/bwear_mat.png')
plt.close()


#boxplot materiale
#myplot = df.boxplot(column='tempo_usura',by='materiale', patch_artist=True)


#myplot.set_ylabel('tempo limite usura [min]')
#myplot.set_title('')
#myplot.set_xlabel('')
#plt.show()
#plt.savefig('/data/metalm/output/bwear_mat.png')
```

## Comparativa di tutte le schede analizzate

```{python boxplot_sk}

materiali = np.unique(df.materiale)

array = []

schedelabel = []

i=0

for materiale in materiali:
  
  df_materiale = df[df.materiale == materiale]
  
  schede = np.unique(df_materiale.sk)
    
  for scheda in schede:
    
    colore = str(colori_materiali.loc[materiale,'colore'])
      
    df_scheda = df_materiale[df_materiale.sk == scheda]

    #print(colore, materiale, scheda)
    
    myplot = df_scheda.boxplot(column='tempo_usura', 
    #patch_artist=True,
    #notch = True,
    color = colore,
    positions = [i],
    rot=90,
    )
    
    array.append(i)
    
    schedelabel.append(scheda)
    
    i=i+1
    
    #print(array)
    
plt.xticks(array, schedelabel)
myplot.set_ylabel('tempo limite usura [min]')
myplot.set_title('')
myplot.set_xlabel('')
plt.show()
plt.savefig('/data/metalm/output/bwear_sk.png')
plt.close()


#boxplot materiale e scheda

#myplot = df.boxplot(column='tempo_usura',by=['materiale','sk'],rot=90, fontsize=10, figsize=(15, 20))
#myplot.set_ylabel('tempo limite usura [min]')
#myplot.set_title('')
#myplot.set_xlabel('')
#plt.show()
#plt.savefig('/data/metalm/output/bwear_sk.png')

```

## Tabella risultati prove usura

```{python tabella_risultati}

#calcolo medie per materiale

materiali=list(np.unique(df.materiale))

dfa = pd.DataFrame()

all_count = []
all_mean = []
all_std = []
all_mat = []
for mat in materiali:
  media=df[df.materiale==mat].tempo_usura.mean()
  std=df[df.materiale==mat].tempo_usura.std()
  count=df[df.materiale==mat].tempo_usura.count()
  
  all_mean.append(media)
  all_std.append(std)
  all_count.append(count)
  all_mat.append(mat)

dfa['materiale'] = all_mat  
dfa['media_wear'] = all_mean
dfa['media_std'] = all_std
dfa['media_cont'] = all_count


L1=float(dfa[dfa.materiale=='F304L1'].media_wear)
dfa['diff_perc1']=((dfa.media_wear-L1)/L1)*100

dfa.to_csv('/data/metalm/output/resumed_usura.txt',sep='|', float_format='%1.2f', index=False)


```

Tabella riassuntiva delle prove a usura

```{r plot_tabella}

tabella <- read_delim('/data/metalm/output/resumed_usura.txt',
                      delim='|')

cas_kable(tabella, caption='resumed data')

```



# UVAV

leggiamo i dati

```{python lettura_opt}

## grouped by materiale
dati = pd.read_pickle('/data/metalm/pickle/opt.pkl')

#togliamo LH dai dati
dati = dati.drop(dati[dati.materiale=='F304LH'].index) 

```

## Comparativa forze assorbite per materiale

```{python UVAV_mat}

all_prove = ['Speed variabile','Feed variabile', 'DOC variabile']


for prova in all_prove:
    
    dfp = dati[dati.tipo_prova == prova].reset_index()
    
    materiali = np.unique(dati.materiale)
    
    plt.figure()
    
    for mat in materiali:
        dfm = dfp[dfp.materiale == mat].reset_index()
        dfg = dfm.groupby('passata').mean()
        dfe = dfm.groupby('passata').std()
        
        colore = str(colori_materiali.loc[mat,'colore'])
        
        if 'Speed' in prova:
            xx = dfg['speed']
            xlab = 'speed [m/min]'
            
        elif 'Feed' in prova:
            xx = dfg['feed']
            xlab = 'feed [mm/rev]'

        elif 'DOC' in prova:
            xx = dfg['doc']
            xlab = 'DOC [mm]'

        else:
            xx = range(0,dfg.shape[0])
            xlab = 'sequence'
            
        err = dfe['Kn']/2
        plt.plot(xx,dfg['Kn'],'-v', label=mat, color = colore)
        plt.fill_between(xx,dfg['Kn']-err,dfg['Kn']+err,alpha=0.3, color= colore)

    plt.legend(loc='best')
    plt.title(prova)
    plt.ylabel('Kn [N/mm2]')
    plt.xlabel(xlab)
    plt.show()
    plt.savefig('/data/metalm/output/'+prova+'.png')

```

## Comparativa forze assorbite delle schede di ogni materiale

```{python UVAV_sk}

## MATERIALI con tutte le schede

all_prove = ['Speed variabile','Feed variabile', 'DOC variabile']

for prova in all_prove:
  
  dati1 = dati[dati.tipo_prova == prova]
  
  materiali = np.unique(dati.materiale)
  
  for mat in materiali:
    
    df0 = dati1[dati1.materiale == mat]
    plt.figure()
    schede = np.unique(df0.sk)
    for scheda in schede:
      
      if 'Speed' in prova:
            xx = 'speed'
            xlab = 'speed [m/min]'
            
      elif 'Feed' in prova:
          xx = 'feed'
          xlab = 'feed [mm/rev]'

      elif 'DOC' in prova:
          xx = 'doc'
          xlab = 'DOC [mm]'
      
      df = df0[df0.sk == scheda]
      df2 = df.groupby('passata').mean()
      df2e = df.groupby('passata').std()
      plt.plot(df2[xx],df2['Kn'],'-v',label=scheda)
      yerr = df2e['Kn']/2
      plt.fill_between(df2[xx],df2['Kn']-yerr,df2['Kn']+yerr,alpha=0.3)

    plt.xlabel(xlab)
    plt.ylabel('Kn [N/mm2]')
    plt.legend(loc = 'best')
    plt.title(mat+'\n'+prova)
    plt.show()
    plt.savefig('/data/metalm/output/'+mat[0:6]+xx+'.png')
  


```

## Comparativa tra tutte le schede analizzate

```{python UVAV_tutte_sk}

prove = ['Speed variabile','Feed variabile', 'DOC variabile']


for prova in prove:
  
  df_prova = dati[dati.tipo_prova == prova]
  
  materiali = np.unique(df_prova.materiale)
  
  plt.figure()
  
  for materiale in materiali:
    
    df_materiale = df_prova[df_prova.materiale == materiale]
    
    schede = np.unique(df_materiale.sk)
    
    for scheda in schede:
      
      df_scheda = df_materiale[df_materiale.sk == scheda]
      medie_sk = df_scheda.groupby('passata').mean()
      sdt_sk = df_scheda.groupby('passata').std()
      
      colore = str(colori_materiali.loc[materiale,'colore'])
      
      medie_sk.columns = medie_sk.columns.str.lower()
      
      
      listaprova = prova.split()
      tipoprova1 = listaprova[0]
      tipoprova = tipoprova1.lower()
     
      prova_col1 = [col for col in medie_sk.columns if tipoprova in col]
      prova_col = prova_col1[0]
      
      plt.plot(medie_sk[prova_col],medie_sk['kn'],'-v',label=scheda, color = colore )
  
  #plt.annotate(colori_materiali.index, labelcolor = colori_materiali.colore)
  
  plt.legend(loc='upper right')
  plt.title(prova)
  plt.ylabel('Kn [N/mm2]')
  plt.show()
  plt.savefig('/data/metalm/output/'+prova+' tutte le sk.png')


```

## Differenze di energia durante le prove UVAV

```{python differenze_energia}

#leggo il file
file_mach = '/data/metalm/pickle/opt.pkl'
df = pd.read_pickle(file_mach)

#droppo F304LH
df0 = df.drop(df[df.materiale=='F304LH'].index,axis=0).reset_index()

#estraggo le features che mi interessano
all_prove = np.unique(df.tipo_prova)
all_colate = np.unique(df.colata)


processed_data = []
for col in all_colate:
    df1 = df0[df0.colata == col].reset_index()
    df1 =df1.drop('level_0',axis=1)
    
    for prova in all_prove:
        if not 'DOE' in prova:
            df2 = df1[df1.tipo_prova==prova].reset_index()
            df2 =df2.drop('level_0',axis=1)
            all_materiali = np.unique(df2.materiale)
            
            for mat in all_materiali:
                
                df3 = df2[df2.materiale==mat].reset_index()
                all_schede = np.unique(df3.sk)
                df3 =df3.drop('level_0',axis=1)
                for scheda in all_schede:
                    df4 = df3[df3.sk==scheda].reset_index()
                    
                    dfg = df4.groupby('passata').mean()
                    
                    processed_data.append([prova,col, mat,scheda,dfg['Kn'].mean()])

#creo il dataframe dai dati preprocessati
data = pd.DataFrame(processed_data,columns=['prova','colata','materiale','sk','Kn_mean'])

# raggruppo i dati per il barplot
dict_comparativa = {}
for prova in all_prove:
    if not 'DOE' in prova:
        confronto = []
    
        dati_i = data[data.prova == prova]
        dd = dati_i.groupby('materiale').mean().Kn_mean
        
        r = len(dd)
        
        for i in range(1,r):
            
            confronto.append((dd[i]-dd[0])/dd[0]*100)
        dict_comparativa[prova] = confronto
    
       
dati = pd.DataFrame(dict_comparativa,index=['F304LO','F304LX'])

#faccio il plot delle differenze di energia
plot = dati.plot.bar(rot=0)

# formatto come 1.2f per rendere leggibile la tabells
dati.update(dati.applymap('{:,.2f}'.format))


# creo la tabella
table = tbl.table(plot, cellText = dati.values.T,
                      cellLoc='center',
                      colLabels = ['',''],
                      rowLabels = ['DOC variabile', 'Feed variabile', 'Speed variabile'],
                      loc ='bottom')
plot.add_table(table)
plot.set_xlabel(xlabel=None)
plot.set_ylabel('delta perc')
plot.set_title('F304L1 \nDifferenze medie percentuali di energia')
plt.show()
plt.savefig('/data/metalm/output/energy_diff.png')

```

# Composizione chimica ossidi

Leggiamo i dati

```{python lettura_ossidi}

path_data = '/data/metalm/pickle/ossidi.pkl'
df = pd.read_pickle(path_data)

df['sk'] = df['sk'].astype(str)
df['col'] = df['col'].astype(str)

df['AL2O3_p']=df.AL2O3_i*100
df['SIO2_p'] = df.SIO2_i*100
df['CAO_p'] = df.CAO_i*100

```

Definiamo la funzione per plottare un grafico a tre variabili

```{python triplotter_ossidi}

def trplt(df,mode,savepic):
  # Scatter Plot ternario
  # mode sono le marche o le schede
  # savepic è il nome di com voglio il png
  
  plt.style.use('seaborn')
  scale = 100
  size = 30
  figure, tax = ternary.figure(scale=scale)
  figure.set_size_inches(15, 15)
  
  colori = ['r','b','g','m','orange','lime','indigo', 'cyan','navy','k','sienna', 'purple', 'teal', 'grey', 'maroon', 'gold', 'hotpink', 'coral', 'tan', 'azure', 'snow', 'salmon', 'olive', 'pink']
  i = 0

  for tt in list(np.unique(df[mode])):
    dfi = df[df[mode] == tt].reset_index()
    # Plot a few different styles with a legend
    dfii = dfi[dfi.kind == 'D'].reset_index()
    points = dfii[['AL2O3_p','SIO2_p','CAO_p']].values
    tax.scatter(points, marker='o',s=size,c=colori[i], label=tt+'| D type')
  
    dfii = dfi[dfi.kind == 'B-C'].reset_index()
    points = dfii[['AL2O3_p','SIO2_p','CAO_p']].values
    tax.scatter(points, marker='s',s=size,c=colori[i], label=tt+'| B-C type')
    
    i += 1
    
  tax.legend(bbox_to_anchor=(0, 1), fontsize=10)
  tax.set_title("inclusion ternary plot", fontsize=25)
  
  # Set Axis labels and Title
  fontsize = 15
  tax.left_axis_label('SiO2 [%wt]', fontsize=fontsize, offset=0.1)
  tax.right_axis_label('CaO [%wt]', fontsize=fontsize, offset=0.1)
  tax.bottom_axis_label('Al2O3 [%wt]', fontsize=fontsize, offset=0.01)
  
  tax.boundary(linewidth=2.0)
  tax.gridlines(multiple=5, color="blue")
  tax.ticks(axis='lbr', linewidth=1, multiple=5)
  tax.clear_matplotlib_ticks()
  tax.get_axes().axis('off')
  
  tax.show()
  tax.savefig ('/data/metalm/output/%s.png' %savepic)
  tax.close()
  plt.style.use('default')

```

## Composizione ossidi per materiale

```{python triplotter_ossidi_mat}

trplt(df,'materiale','mat')

```

## Composizione ossidi per colata

```{python triplotter_ossidi_col}

trplt(df,'col','col')

```

## Composizione ossidi per scheda

```{python triplotter_ossidi_sk}

trplt(df,'sk','schede')

```


# tempo usura VS variabili risposta

## esame delle variabili (input VS output)

leggiamo i dati dal file .pkl

```{python loading_dati}
import pandas as pd

df0 = pd.read_pickle('/data/metalm/pickle/dbmm.pkl')
df0 = df0[df0.materiale != 'F304LH']

features0 = ['Fz_iniziale','Fz_peak','dF_perc', 'C', 'S', 'P', 'Si', 'Mn', 'Cr', 'Ni', 'Mo','Cu', 'Sn', 'Al', 'V', 'W', 'Co',
'Ti', 'Nb', 'B', 'Ca', 'N2', 'O2 ','k2_oa', 'k2_os', 'k2_og', 'rm', 'rp02']

features = ['tempo_usura','C', 'S', 'P', 'Si', 'Mn', 'Cr', 'Ni', 'Mo','Cu', 'Sn', 'Al', 'V', 'W', 'Co',
'Ti', 'Nb', 'B', 'Ca', 'N2', 'O2 ']
       
df = df0[features]
```
## tempo_usura VS variabili risposta

```{r correlazioni,results='asis'}
dati <- py$df0
myfeatures <- py$features0

response <- 'tempo_usura'

vars <- dati %>% 
  select(myfeatures) %>%
  names()
n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  pl <- ggplot(dati) +
    geom_point(aes(!!sym(var_i),!!sym(response))) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'red',
      se = FALSE,
      method = 'loess'
    ) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'green',
      se = FALSE,
      method = 'lm'
    )
  title <- paste('## Variabile: ', var_i)
  cat(title, '\n')
  print(pl)
  cat('<p>') #necessario per terminare paragrafo
}
```
